refactor(apim): modularize API Management service handler
Extract ARM management logic to ApiManagementService.
Extract gateway runtime logic to ApiManagementGateway.
Extract HTTP proxying to ApiManagementProxy.
Centralize data stores in ApiManagementStore.
Improve separation of concerns and maintainability.
